# swagger.json
{"openapi":"3.0.0","paths":{"/v1":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/v1/auth/admin/local/sign-up":{"post":{"operationId":"AuthController_signUpLocal","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpWithEmailDto"}}}},"responses":{"201":{"description":"Returns new user & business record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSuccessResponse"}}}}},"tags":["auth"]}},"/v1/auth/admin/local/sign-in":{"post":{"operationId":"AuthController_signInLocal","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInWithEmailDto"}}}},"responses":{"201":{"description":"Returns existing user record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthSuccessResponse"}}}}},"tags":["auth"]}},"/v1/auth/admin/local/sign-out":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":"Revokes tokens"}},"tags":["auth"],"security":[{"bearer":[]}]}},"/v1/auth/admin/local/refresh":{"post":{"operationId":"AuthController_refreshTokens","parameters":[],"responses":{"201":{"description":"Returns refreshed token"}},"tags":["auth"]}},"/v1/business/{businessId}":{"put":{"operationId":"BusinessController_updateBusiness","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBusinessDto"}}}},"responses":{"201":{"description":"Updates business record"}},"tags":["business"],"security":[{"bearer":[]}]},"get":{"operationId":"BusinessController_getBusiness","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Returns business record"}},"tags":["business"],"security":[{"bearer":[]}]}},"/v1/business/{businessId}/admins":{"get":{"operationId":"BusinessController_getBusinessAdmins","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Returns business admins"}},"tags":["business"],"security":[{"bearer":[]}]}},"/v1/business/{businessId}/services":{"get":{"operationId":"BusinessController_getBusinessServices","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Returns business services"}},"tags":["business"],"security":[{"bearer":[]}]}},"/v1/business/{businessId}/categories":{"get":{"operationId":"BusinessController_getBusinessCategories","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Returns business categories"}},"tags":["business"],"security":[{"bearer":[]}]}},"/v1/business/{businessId}/hours":{"get":{"operationId":"BusinessController_getBusinessHours","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Returns business hours"}},"tags":["business"],"security":[{"bearer":[]}]}},"/v1/business/{businessId}/events":{"get":{"operationId":"BusinessController_getBusinessEvents","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Returns business events"}},"tags":["business"],"security":[{"bearer":[]}]}},"/v1/business/{businessId}/clients":{"get":{"operationId":"BusinessController_getBusinessClients","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Returns business clients"}},"tags":["business"],"security":[{"bearer":[]}]},"post":{"operationId":"BusinessController_createBusinessClients","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"201":{"description":"Returns created"}},"tags":["business"],"security":[{"bearer":[]}]}},"/v1/business/{businessId}/staff":{"get":{"operationId":"BusinessController_getBusinessStaffByIds","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Returns business staff"}},"tags":["business"],"security":[{"bearer":[]}]},"post":{"operationId":"BusinessController_createBusinessStaff","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"201":{"description":"Returns business staff"}},"tags":["business"],"security":[{"bearer":[]}]}},"/v1/business/registration/complete":{"put":{"operationId":"BusinessController_registerBusiness","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterBusinessDto"}}}},"responses":{"201":{"description":"Register Business"}},"tags":["business"],"security":[{"bearer":[]}]}},"/v1/shortwaits/admin/mobile":{"get":{"operationId":"ShortwaitsController_getMobileDefaultData","parameters":[],"responses":{"201":{"description":"Returns default data for admin mobile","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Shortwaits"}}}}},"tags":["shortwaits"]}},"/v1/events/business/summary/{businessId}":{"get":{"operationId":"EventsController_getEventsSummaryByBusiness","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Get a summary of all events in a business (business is NOT a user!!!)"}},"tags":["events"],"security":[{"bearer":[]}]}},"/v1/events/business/{businessId}":{"get":{"operationId":"EventsController_getEventsByBusiness","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"date","required":false,"in":"query","schema":{"default":"2023-07-29T08:10:44.918Z","type":"string"}},{"name":"filterBy","required":false,"in":"query","schema":{"default":"year","enum":["day","month","year"],"type":"string"}}],"responses":{"201":{"description":"Get all events in a business (business is NOT a user!!!)"}},"tags":["events"],"security":[{"bearer":[]}]},"post":{"operationId":"EventsController_createEventByBusiness","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventsDto"}}}},"responses":{"201":{"description":"Create new event record by Business User (not client user or business)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Events"}}}}},"tags":["events"],"security":[{"bearer":[]}]},"put":{"operationId":"EventsController_updateEventByBusiness","parameters":[{"name":"businessId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Create new event record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventsDto"}}}}},"tags":["events"],"security":[{"bearer":[]}]}},"/v1/events":{"put":{"operationId":"EventsController_updateEventByUser","parameters":[],"responses":{"201":{"description":"Create new event record","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventsDto"}}}}},"tags":["events"],"security":[{"bearer":[]}]}},"/v1/events/delete/{eventId}":{"put":{"operationId":"EventsController_deleteEvent","parameters":[{"name":"eventId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Event deleted successfully."}},"tags":["events"],"security":[{"bearer":[]}]}},"/v1/events/delete":{"put":{"operationId":"EventsController_deleteEvents","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"201":{"description":"Events deleted successfully."}},"tags":["events"],"security":[{"bearer":[]}]}},"/v1/services":{"post":{"operationId":"ServicesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateServiceDto"}}}},"responses":{"201":{"description":""}},"tags":["services"],"security":[{"bearer":[]}]},"get":{"operationId":"ServicesController_getServices","parameters":[{"name":"businessId","required":false,"in":"query","schema":{"format":"string","type":"string"}},{"name":"serviceId","required":false,"in":"query","schema":{"format":"string","type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":0,"maximum":10000,"exclusiveMaximum":true,"exclusiveMinimum":true,"title":"Page","format":"int32","default":0,"type":"number"}},{"name":"_sortBy","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetServicesDto"}}}},"responses":{"200":{"description":""}},"tags":["services"],"security":[{"bearer":[]}]}},"/v1/services/{serviceId}":{"put":{"operationId":"ServicesController_update","parameters":[{"name":"serviceId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateServiceDto"}}}},"responses":{"200":{"description":""}},"tags":["services"],"security":[{"bearer":[]}]},"patch":{"operationId":"ServicesController_updateSome","parameters":[{"name":"serviceId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateServiceDto"}}}},"responses":{"200":{"description":""}},"tags":["services"],"security":[{"bearer":[]}]},"delete":{"operationId":"ServicesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["services"],"security":[{"bearer":[]}]}},"/v1/categories":{"get":{"operationId":"CategoriesController_getAllCategories","parameters":[],"responses":{"201":{"description":"Returns all categories record"}},"tags":["categories"],"security":[{"bearer":[]}]}},"/v1/categories/{id}":{"get":{"operationId":"CategoriesController_getCategory","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"Returns category record"}},"tags":["categories"],"security":[{"bearer":[]}]}},"/v1/client-user/multiple":{"post":{"operationId":"ClientUserController_getMultipleUsers","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"200":{"description":""}},"tags":["client-user"],"security":[{"bearer":[]}]}},"/v1/client-user":{"post":{"operationId":"ClientUserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClientUserDto"}}}},"responses":{"201":{"description":""}},"tags":["client-user"],"security":[{"bearer":[]}]}},"/v1/business-user/multiple":{"post":{"operationId":"BusinessUsersController_getMultipleUsers","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}},"responses":{"200":{"description":""}},"tags":["business-user"],"security":[{"bearer":[]}]}},"/v1/business-user":{"post":{"operationId":"BusinessUsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBusinessUserDto"}}}},"responses":{"201":{"description":""}},"tags":["business-user"],"security":[{"bearer":[]}]}}},"info":{"title":"Shortwaits Admin - API","description":"Shortwaits's successful response: <br>\n<code>{data: [\n{ id: 1, name: \"Item 1\" },\n{ id: 2, name: \"Item 2\" },\n{ id: 3, name: \"Item 3\" },\n],<br>\nstatusCode: 200,<br>\nmeta: {\npagination: {\ntotalItems: 10,\ntotalPages: 2,\ncurrentPage: 1,\nitemsPerPage: 3,\n},\n},\n}<code>\n","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://127.0.0.1:8080","description":"Dev server"},{"url":"https://api.shortwaits.com","description":"Production server"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignUpWithEmailDto":{"type":"object","properties":{"username":{"type":"string","example":"sw"},"email":{"type":"string","example":"sw@sw.com"},"password":{"type":"string","example":"Shortwaits123"}},"required":["username","email","password"]},"AuthSuccessResponse":{"type":"object","properties":{"auth":{"type":"object"},"attributes":{"type":"object"}},"required":["auth","attributes"]},"SignInWithEmailDto":{"type":"object","properties":{"username":{"type":"string","example":"sw123"},"email":{"type":"string","example":"sw@sw.com"},"password":{"type":"string","example":"Shortwaits123"}},"required":["username","email","password"]},"UpdateBusinessDto":{"type":"object","properties":{"country":{"type":"string"},"phone1":{"type":"string"},"longName":{"type":"string"},"shortName":{"type":"string"},"description":{"type":"string"},"isRegistrationCompleted":{"type":"boolean"},"staff":{"type":"array","items":{"type":"string"}},"categories":{"type":"array","items":{"type":"string"}},"services":{"type":"array","items":{"type":"string"}}},"required":["country","phone1","longName","shortName","description","isRegistrationCompleted","staff","categories","services"]},"RegisterBusinessDto":{"type":"object","properties":{"shortName":{"type":"string"},"description":{"type":"string"},"isRegistrationCompleted":{"type":"boolean"},"staff":{"type":"array","items":{"type":"string"}},"categories":{"type":"array","items":{"type":"string"}},"services":{"type":"array","items":{"type":"string"}}},"required":["shortName","description","isRegistrationCompleted","staff","categories","services"]},"Shortwaits":{"type":"object","properties":{"banners":{"type":"array","items":{"type":"string"}},"short_id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"links":{"type":"array","items":{"type":"string"}},"suggestedLang":{"type":"string"},"blackList":{"type":"array","items":{"type":"string"}},"timeZones":{"type":"array","items":{"type":"string"}},"categories":{"type":"array","items":{"type":"string"}},"sampleBusinessData":{"type":"object"},"serviceColors":{"type":"object"}},"required":["banners","short_id","name","description","links","suggestedLang","blackList","timeZones","categories","sampleBusinessData","serviceColors"]},"LocationDto":{"type":"object","properties":{}},"CreateEventsDto":{"type":"object","properties":{"expectedEndTime":{"type":"string"},"registrationDeadlineTime":{"type":"string"},"paymentMethod":{"type":"object"},"participantsIds":{"type":"array","items":{"type":"string"}},"leadClientId":{"type":"string"},"urls":{"type":"array","items":{"type":"string"}},"location":{"$ref":"#/components/schemas/LocationDto"},"attendeeLimit":{"type":"number"},"registrationDeadline":{"format":"date-time","type":"string"},"registrationFee":{"type":"number"},"serviceId":{"type":"string"},"staffIds":{"type":"array","items":{"type":"string"}},"clientsIds":{"type":"array","items":{"type":"string"}},"hasNoDuration":{"type":"boolean"},"eventImage":{"type":"string"},"businessId":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"features":{"type":"array","items":{"type":"string"}},"durationInMin":{"type":"number"},"startTime":{"type":"string"},"endTime":{"type":"string"},"endTimeExpected":{"type":"string"},"priceExpected":{"type":"number"},"isGroupEvent":{"type":"boolean"},"repeat":{"type":"boolean"},"payment":{"type":"object"},"notes":{"type":"string"},"labels":{"type":"object"}},"required":["expectedEndTime","paymentMethod","participantsIds","serviceId","staffIds","clientsIds","businessId","name","description","startTime","priceExpected"]},"EventLocation":{"type":"object","properties":{}},"Events":{"type":"object","properties":{"expectedEndTime":{"format":"date-time","type":"string"},"registrationDeadlineTime":{"format":"date-time","type":"string"},"paymentMethod":{"type":"object"},"participantsIds":{"type":"array","items":{"type":"string"}},"leadClientId":{"type":"object"},"urls":{"type":"array","items":{"type":"string"}},"location":{"$ref":"#/components/schemas/EventLocation"},"attendeeLimit":{"type":"number"},"registrationFee":{"type":"number"},"serviceId":{"type":"object"},"staffIds":{"type":"array","items":{"type":"string"}},"clientsIds":{"type":"array","items":{"type":"string"}},"hasNoDuration":{"type":"boolean"},"leadClientName":{"type":"string"},"eventImage":{"type":"string"},"businessId":{"type":"object"},"name":{"type":"string"},"description":{"type":"string"},"createdBy":{"type":"object"},"updatedBy":{"type":"object"},"staff":{"type":"array","items":{"type":"string"}},"clients":{"type":"array","items":{"type":"string"}},"features":{"type":"array","items":{"type":"string"}},"status":{"type":"object"},"durationInMin":{"type":"number"},"startTime":{"format":"date-time","type":"string"},"endTime":{"format":"date-time","type":"string"},"endTimeExpected":{"format":"date-time","type":"string"},"priceExpected":{"type":"number"},"priceFinal":{"type":"number"},"canceled":{"type":"boolean"},"cancellationReason":{"type":"string"},"isGroupEvent":{"type":"boolean"},"repeat":{"type":"boolean"},"payment":{"type":"object"},"notes":{"type":"string"},"labels":{"type":"object"},"deleted":{"type":"boolean"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["expectedEndTime","registrationDeadlineTime","paymentMethod","participantsIds","leadClientId","urls","location","attendeeLimit","registrationFee","serviceId","staffIds","clientsIds","hasNoDuration","leadClientName","eventImage","businessId","name","description","createdBy","updatedBy","staff","clients","features","status","durationInMin","startTime","endTime","endTimeExpected","priceExpected","priceFinal","canceled","cancellationReason","isGroupEvent","repeat","payment","notes","labels","deleted","createdAt","updatedAt"]},"CreateServiceDto":{"type":"object","properties":{}},"GetServicesDto":{"type":"object","properties":{"services":{"type":"array","items":{"type":"string","format":"array"}}}},"UpdateServiceDto":{"type":"object","properties":{}},"CreateClientUserDto":{"type":"object","properties":{"clientType":{"type":"string"},"username":{"type":"string"},"alias":{"type":"string"},"givenName":{"type":"string"},"familyName":{"type":"string"},"displayName":{"type":"string"},"middleName":{"type":"string"},"accountImageUrl":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"locale":{"type":"object"},"phoneNumbers":{"type":"array","items":{"type":"string"}},"imAddresses":{"type":"array","items":{"type":"string"}},"addresses":{"type":"array","items":{"type":"string"}},"socialAccounts":{"type":"array","items":{"type":"string"}},"desiredCurrencies":{"type":"array","items":{"type":"string"}}},"required":["clientType","username","alias","givenName","familyName","displayName","middleName","accountImageUrl","email","password","locale","phoneNumbers","imAddresses","addresses","socialAccounts","desiredCurrencies"]},"CreateBusinessUserDto":{"type":"object","properties":{"birthday":{"type":"object"},"registrationState":{"type":"object"},"roleId":{"type":"string"},"clientType":{"type":"string"},"username":{"type":"string"},"alias":{"type":"string"},"givenName":{"type":"string"},"familyName":{"type":"string"},"displayName":{"type":"string"},"middleName":{"type":"string"},"accountImageUrl":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"locale":{"type":"object"},"phoneNumbers":{"type":"array","items":{"type":"string"}},"imAddresses":{"type":"array","items":{"type":"string"}},"addresses":{"type":"array","items":{"type":"string"}},"socialAccounts":{"type":"array","items":{"type":"string"}},"desiredCurrencies":{"type":"array","items":{"type":"string"}}},"required":["birthday","registrationState","roleId","clientType","username","alias","givenName","familyName","displayName","middleName","accountImageUrl","email","password","locale","phoneNumbers","imAddresses","addresses","socialAccounts","desiredCurrencies"]}}}}